rag-backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9

# OpenAI and AI dependencies
openai==1.54.0
tiktoken==0.8.0

# Qdrant vector database
qdrant-client==1.11.0

# Text processing and RAG
langchain==0.3.0
langchain-text-splitters==0.3.0
python-frontmatter==1.1.0
markdown==3.7

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
tenacity>=8.1.0,<9.0.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Development
black==24.8.0
ruff==0.7.0
